support for 'from cryptotools import *' wildcard import
